# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DataX Enterprise Studio API
  version: 1.2.0
  license:
    name: License pending repository-owner decision
    url: https://github.com/xiaoli2hust/datax/blob/main/README.md
  description: |
    V1 单组织、多项目、安全的一次性离线全量复制控制 API。
    仅支持 MySQL 8/PostgreSQL 15、人工触发、源端静默确认、预存在且实测为空的目标表和 insert-only-once。
    不提供调度、DAG、告警、AI、CDC、自由 SQL、Transformer 或插件安装接口。
    交付形态为 Windows 11 x64 本地工作站；业务 API 仅经 web 反向代理绑定本机
    127.0.0.1:17860，api、worker 和 postgres 不暴露宿主端口。
servers:
  - url: http://127.0.0.1:17860/api/v1
    description: Windows 本地工作站唯一支持入口；不得替换为 0.0.0.0、LAN IP 或公网地址
...
security:
  - bearerAuth: []
paths:
...
  /executions/{execution_id}/recovery:
    parameters:
      - $ref: '#/components/parameters/ExecutionId'
    get:
      tags: [Executions]
      operationId: getExecutionRecoveryGate
      summary: 获取失败执行的人工恢复门禁
      description: 仅已领取且未以 SUCCEEDED/CONFIRMED/PASSED 结束的 Execution 才有 gate；未领取即取消的 CANCELED 不适用。
      responses:
        '200':
          description: 恢复门禁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryGate'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    post:
      tags: [Executions]
      operationId: submitExecutionRemediation
      summary: 提交处置确认（remediation confirmation）并排队 RecoveryProbe
      description: |
        已经领取且 TargetCopyLock 已从 RESERVED 转为 ACTIVE，随后进入 RECOVERY_REQUIRED 的
        FAILED、TIMED_OUT、CANCELED 或 LOST 执行再次执行前需要该门禁；未领取即取消的
        CANCELED 没有 Attempt/fence/gate，RESERVED 锁已释放，
        recovery 接口返回 RECOVERY_GATE_NOT_APPLICABLE。
        即使 data_effect=NONE 也必须如实提交处置确认；无需清理时使用
        action=NO_CLEANUP_REQUIRED、cleanup_performed=false 并说明 reason。任何声明都不能
        跳过 Worker 空表复检。
        API 只持久化 REMEDIATION_SUBMITTED；Worker 取得目标锁后生成 target_empty_evidence，
        通过才写 VERIFIED。平台不代替目标表所有者执行 truncate、delete、drop 或建表。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemediationConfirmation'
            example:
              action: NO_CLEANUP_REQUIRED
              cleanup_performed: false
              reason: DataX 启动前即失败，目标所有者确认无需执行清理；仍等待平台空表复检
              confirmed_at: '2026-07-30T03:00:00Z'
      responses:
        '202':
          description: 已提交处置确认并创建独立 RecoveryProbe 事实
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoverySubmission'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /recovery-probes/{recovery_probe_id}:
    parameters:
      - name: recovery_probe_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [Executions]
      operationId: getRecoveryProbe
      summary: 获取恢复空表探针状态
      description: RecoveryProbe 是独立 PostgreSQL 事实队列，使用自身 Attempt、lease 和 fence；不启动 DataX。
      responses:
        '200':
          description: RecoveryProbe 安全摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecoveryProbe'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /executions/{execution_id}/rerun:
    parameters:
      - $ref: '#/components/parameters/ExecutionId'
    post:
      tags: [Executions]
      operationId: rerunExecution
      summary: 恢复后使用原 JobVersion 再次执行
      description: |
        只允许对已经领取、存在 RecoveryGate 且状态为 FAILED、TIMED_OUT、CANCELED 或
        LOST 的原执行恢复。未领取即取消的 CANCELED 不走该路径。
        原执行无论 data_effect 当前判断为何，都必须完成必要外部处置并如实提交处置确认；无需
        清理时不得虚假声称清理，但仍须关联 RecoveryGate
        已由 Worker 空表复检为 VERIFIED；状态判断不能代替复检。
        客户端还须再次确认源端静默，并提交有版本、有限有效期且未撤回的目标排他声明；
        Worker 在新执行启动前仍会取得同目标活动锁并再次实测目标为空，oracle 目标快照结束时间不得晚于 valid_until。
        不会自动选择最新 JobVersion，也不提供自动重试。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecutionRerunCreate'
            example:
              source_quiescence_confirmation:
                confirmed: true
                confirmed_at: '2026-07-30T03:10:00Z'
                note: 已再次暂停源表写入
              target_exclusivity_confirmation:
                statement_version: '1.0'
                confirmed: true
                confirmed_at: '2026-07-30T03:10:05Z'
                valid_until: '2026-07-30T07:10:05Z'
                responsible_party: OPERATOR
                note: 已再次冻结目标表平台外 DML/DDL
              recovery_gate_id: 44444444-4444-4444-8444-444444444444
      responses:
        '202':
          description: 新 Execution
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/CapacityAdmissionBlocked'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: des_refresh
      description: V1 loopback HTTP 使用 HttpOnly、SameSite=Strict、Path=/api/v1/auth 且无 Domain；若未来改为受信 loopback HTTPS 必须增加 Secure。
  headers:
...
    IdempotencyReplayed:
      description: true 表示返回 24 小时内首次请求结果
      schema:
        type: boolean
  parameters:
...
    ExecutionId:
      name: execution_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 资源不存在或对当前主体不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 版本、幂等、状态或唯一约束冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    CapacityAdmissionBlocked:
      description: 普通新 Execution 因持久容量准入门禁被拒绝
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/CapacityAdmissionProblem'
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    NullableTimestamp:
      oneOf:
        - $ref: '#/components/schemas/Timestamp'
        - type: 'null'
    NullableUuid:
      oneOf:
        - $ref: '#/components/schemas/Uuid'
        - type: 'null'
...
    ProcessState:
      type: string
      enum:
        - QUEUED
        - STARTING
        - RUNNING
        - VERIFYING
        - SUCCEEDED
        - FAILED
        - TIMED_OUT
        - CANCEL_REQUESTED
        - CANCELED
        - LOST
    DataEffect:
      type: string
      description: 目标端影响是否已测得，独立于 DataX 进程状态和内容正确性；CONFIRMED 表示影响已测得，测得值可以是 0 行，绝不等于内容核验通过。空源成功仍为 SUCCEEDED/CONFIRMED/PASSED。
      enum: [NONE, POSSIBLE, CONFIRMED, UNKNOWN]
...
    SourceQuiescenceConfirmation:
      type: object
      additionalProperties: false
      description: 操作员声明源表从运行前检查开始到独立 oracle 完成始终停止业务写入；actor 来自认证上下文，V1 只定义 confirmed/confirmed_at/note，不虚构过期或撤回生命周期。核验前后指纹不变是必要证据，但该声明和指纹都不等于数据库一致性快照，也不能证明未报告或瞬时后回滚的写入从未发生。
      required: [confirmed, confirmed_at]
      properties:
        confirmed:
          const: true
        confirmed_at:
          $ref: '#/components/schemas/Timestamp'
        note:
          type: [string, 'null']
          maxLength: 500
...
    RemediationConfirmation:
      description: 非成功执行后由 Operator/目标所有者对必要外部处置作出的处置确认（remediation confirmation）；无需清理时必须如实记录理由。无论是否执行清理，都不能替代 Worker 的目标空表复检。
      oneOf:
        - type: object
          additionalProperties: false
          title: 已清理或重建目标
          required: [action, cleanup_performed, reason, confirmed_at]
          properties:
            action:
              type: string
              enum: [CLEANED_TARGET, RECREATED_TARGET]
            cleanup_performed:
              const: true
            reason:
              type: string
              minLength: 1
              maxLength: 1000
            confirmed_at:
              $ref: '#/components/schemas/Timestamp'
        - type: object
          additionalProperties: false
          title: 经判断无需清理
          required: [action, cleanup_performed, reason, confirmed_at]
          properties:
            action:
              const: NO_CLEANUP_REQUIRED
            cleanup_performed:
              const: false
            reason:
              type: string
              minLength: 1
              maxLength: 1000
            confirmed_at:
              $ref: '#/components/schemas/Timestamp'
        - type: object
          additionalProperties: false
          title: 其他外部处置
          required: [action, cleanup_performed, reason, confirmed_at]
          properties:
            action:
              const: OTHER_EXTERNAL_ACTION
            cleanup_performed:
              type: boolean
            reason:
              type: string
              minLength: 1
              maxLength: 1000
            confirmed_at:
              $ref: '#/components/schemas/Timestamp'
...
    Execution:
      type: object
      additionalProperties: false
      required:
        - id
        - project_id
        - job_id
        - job_version_id
        - rerun_of_execution_id
        - trigger_type
        - requested_by
        - process_state
        - data_effect
        - verification_state
        - target_exclusivity_confirmation
        - target_exclusivity_status
        - target_exclusivity_revoked_at
        - target_exclusivity_revocation_reason
        - source_datasource_revision_id
        - target_datasource_revision_id
        - source_endpoint_policy_revision_id
        - target_endpoint_policy_revision_id
        - target_namespace_id
        - target_copy_lock
        - source_secret_version
        - target_secret_version
        - source_secret_envelope_id
        - target_secret_envelope_id
        - source_connection_evidence_id
        - target_connection_evidence_id
        - capacity_profile
        - service_reservation_seconds
        - queue_eligibility_state
        - queue_block_reason
        - queue_state_changed_at
        - eligible_wait_milliseconds
        - log_incomplete
        - log_raw_received_bytes
        - log_redacted_received_bytes
        - log_stored_bytes
        - log_dropped_bytes
        - resolved_config_hash
        - runtime_snapshot
        - verification_summary
        - queued_at
        - started_at
        - finished_at
        - exit_code
        - failure_code
        - failure_message
        - summary_parse_status
        - run_summary
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        project_id:
          $ref: '#/components/schemas/Uuid'
        job_id:
          $ref: '#/components/schemas/Uuid'
        job_version_id:
          $ref: '#/components/schemas/Uuid'
        rerun_of_execution_id:
          $ref: '#/components/schemas/NullableUuid'
        trigger_type:
          const: MANUAL
        requested_by:
          $ref: '#/components/schemas/Uuid'
        process_state:
          $ref: '#/components/schemas/ProcessState'
        data_effect:
          $ref: '#/components/schemas/DataEffect'
        verification_state:
          $ref: '#/components/schemas/VerificationState'
        target_exclusivity_confirmation:
          $ref: '#/components/schemas/TargetExclusivityConfirmation'
          description: 创建 Execution 时固化的人工目标排他前提；实际 actor、接受时间与摘要在运行快照和审计事件中固定。
        target_exclusivity_status:
          type: string
          enum: [ACTIVE, REVOKED, EXPIRED]
          description: 最近一次执行状态推进或 oracle 结论时评估并持久化的声明状态；终态后不因墙上时钟流逝改写历史结论。
        target_exclusivity_revoked_at:
          $ref: '#/components/schemas/NullableTimestamp'
        target_exclusivity_revocation_reason:
          type: [string, 'null']
          enum:
            - OPERATOR_REVOKED
            - DBA_REVOKED
            - EXTERNAL_DML_DDL_REPORTED
            - CHANGE_FREEZE_BROKEN
            - VALIDITY_WINDOW_EXPIRED
            - null
        source_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        target_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        source_endpoint_policy_revision_id:
          $ref: '#/components/schemas/Uuid'
          description: 创建 Execution 时从不可变 JobVersion 固定；即使尚未领取也不得为 null。
        target_endpoint_policy_revision_id:
          $ref: '#/components/schemas/Uuid'
          description: 创建 Execution 时从不可变 JobVersion 固定；即使尚未领取也不得为 null。
        target_namespace_id:
          $ref: '#/components/schemas/Uuid'
        target_copy_lock:
          $ref: '#/components/schemas/TargetCopyLockSummary'
        source_secret_version:
          type: [integer, 'null']
          minimum: 1
          description: Worker 固定的安全版本号；不返回 secret ID、密文或密钥材料。
        target_secret_version:
          type: [integer, 'null']
          minimum: 1
          description: Worker 固定的安全版本号；不返回 secret ID、密文或密钥材料。
        source_secret_envelope_id:
          $ref: '#/components/schemas/NullableUuid'
        target_secret_envelope_id:
          $ref: '#/components/schemas/NullableUuid'
        source_connection_evidence_id:
          $ref: '#/components/schemas/NullableUuid'
        target_connection_evidence_id:
          $ref: '#/components/schemas/NullableUuid'
        capacity_profile:
          type: string
          enum: [LIGHTWEIGHT, LARGE]
        service_reservation_seconds:
          type: integer
          enum: [360, 3600]
        queue_eligibility_state:
          type: string
          enum: [ELIGIBLE, BLOCKED]
        queue_block_reason:
          type: [string, 'null']
          maxLength: 64
        queue_state_changed_at:
          $ref: '#/components/schemas/Timestamp'
        eligible_wait_milliseconds:
          type: integer
          minimum: 0
        log_incomplete:
          type: boolean
        log_raw_received_bytes:
          type: integer
          minimum: 0
        log_redacted_received_bytes:
          type: integer
          minimum: 0
        log_stored_bytes:
          type: integer
          minimum: 0
        log_dropped_bytes:
          type: integer
          minimum: 0
        resolved_config_hash:
          oneOf:
            - $ref: '#/components/schemas/Sha256'
            - type: 'null'
          description: 注入凭据后的实际配置只计算哈希，完整配置不持久化、不返回。
        runtime_snapshot:
          oneOf:
            - $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
            - type: 'null'
        verification_summary:
          oneOf:
            - $ref: '#/components/schemas/VerificationSummary'
            - type: 'null'
        queued_at:
          $ref: '#/components/schemas/Timestamp'
        started_at:
          $ref: '#/components/schemas/NullableTimestamp'
        finished_at:
          $ref: '#/components/schemas/NullableTimestamp'
        exit_code:
          type: [integer, 'null']
        failure_code:
          type: [string, 'null']
          maxLength: 64
        failure_message:
          type: [string, 'null']
          maxLength: 2000
        summary_parse_status:
          type: string
          enum: [PENDING, SUCCEEDED, FAILED]
        run_summary:
          oneOf:
            - $ref: '#/components/schemas/RunSummary'
            - type: 'null'
      allOf:
        - if:
            properties:
              process_state:
                const: QUEUED
            required: [process_state]
          then:
            properties:
              target_copy_lock:
                properties:
                  state:
                    const: RESERVED
                  activated_at:
                    type: 'null'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: 'null'
              runtime_snapshot:
                type: 'null'
              started_at:
                type: 'null'
              finished_at:
                type: 'null'
        - if:
            properties:
              process_state:
                const: CANCELED
              target_copy_lock:
                properties:
                  state:
                    const: RELEASED
                  activated_at:
                    type: 'null'
                required: [state, activated_at]
            required: [process_state, target_copy_lock]
          then:
            description: RELEASED 且从未 activated 的目标锁是未领取取消的机器判据；不能再以 runtime_snapshot=null 单独判断，因为已领取但 preflight 尚未完成的 STARTING Attempt 也允许为 null。
            properties:
              runtime_snapshot:
                type: 'null'
              data_effect:
                const: NONE
              verification_state:
                const: NOT_STARTED
              target_copy_lock:
                properties:
                  state:
                    const: RELEASED
                  activated_at:
                    type: 'null'
                  released_at:
                    $ref: '#/components/schemas/Timestamp'
                  fence_epoch:
                    type: 'null'
              started_at:
                type: 'null'
              finished_at:
                $ref: '#/components/schemas/Timestamp'
              verification_summary:
                type: 'null'
        - if:
            properties:
              process_state:
                enum: [STARTING, CANCEL_REQUESTED]
            required: [process_state]
          then:
            description: 这些状态只可能在领取事务提交后出现；领取事务原子生成 Attempt/fence 并激活目标锁。真实 preflight 完成前 runtime_snapshot 允许为 null，完成后一次性固定且不可覆盖。
            properties:
              runtime_snapshot:
                oneOf:
                  - $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
                  - type: 'null'
              target_copy_lock:
                properties:
                  state:
                    const: ACTIVE
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: integer
                    minimum: 1
        - if:
            properties:
              process_state:
                enum: [RUNNING, VERIFYING]
            required: [process_state]
          then:
            description: STARTING 只有在真实 preflight 证据已完整固化后才能进入这些状态。
            properties:
              runtime_snapshot:
                $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
              target_copy_lock:
                properties:
                  state:
                    const: ACTIVE
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: integer
                    minimum: 1
        - if:
            properties:
              process_state:
                enum: [FAILED, LOST]
            required: [process_state]
          then:
            description: 这些非成功终态可能发生在 preflight 完成前或 DataX 启动后；两者都来自已提交的领取事务并保留恢复门禁，不能以 runtime_snapshot 是否为 null 判断是否领取。
            properties:
              runtime_snapshot:
                oneOf:
                  - $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
                  - type: 'null'
              target_copy_lock:
                properties:
                  state:
                    const: RECOVERY_REQUIRED
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: integer
                    minimum: 1
        - if:
            properties:
              process_state:
                const: TIMED_OUT
            required: [process_state]
          then:
            description: TIMED_OUT 只能发生在 DataX 已启动后，因此必须已有完整 preflight 运行快照并保留恢复门禁。
            properties:
              runtime_snapshot:
                $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
              target_copy_lock:
                properties:
                  state:
                    const: RECOVERY_REQUIRED
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: integer
                    minimum: 1
        - if:
            properties:
              process_state:
                const: CANCELED
              target_copy_lock:
                properties:
                  state:
                    const: RECOVERY_REQUIRED
                required: [state]
            required: [process_state, target_copy_lock]
          then:
            description: RECOVERY_REQUIRED 表示已领取取消；取消可发生在 preflight 完成前，因此 runtime_snapshot 可为 null，但活动 fence/Attempt 事实必须保留。
            properties:
              runtime_snapshot:
                oneOf:
                  - $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
                  - type: 'null'
              target_copy_lock:
                properties:
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    type: 'null'
                  fence_epoch:
                    type: integer
                    minimum: 1
        - if:
            properties:
              target_exclusivity_status:
                const: ACTIVE
            required: [target_exclusivity_status]
          then:
            properties:
              target_exclusivity_revoked_at:
                type: 'null'
              target_exclusivity_revocation_reason:
                type: 'null'
        - if:
            properties:
              target_exclusivity_status:
                const: REVOKED
            required: [target_exclusivity_status]
          then:
            properties:
              target_exclusivity_revoked_at:
                $ref: '#/components/schemas/Timestamp'
              target_exclusivity_revocation_reason:
                type: string
                enum:
                  - OPERATOR_REVOKED
                  - DBA_REVOKED
                  - EXTERNAL_DML_DDL_REPORTED
                  - CHANGE_FREEZE_BROKEN
        - if:
            properties:
              target_exclusivity_status:
                const: EXPIRED
            required: [target_exclusivity_status]
          then:
            properties:
              target_exclusivity_revoked_at:
                type: 'null'
              target_exclusivity_revocation_reason:
                const: VALIDITY_WINDOW_EXPIRED
        - if:
            properties:
              process_state:
                const: SUCCEEDED
            required: [process_state]
          then:
            properties:
              runtime_snapshot:
                $ref: '#/components/schemas/ExecutionRuntimeSnapshot'
              data_effect:
                const: CONFIRMED
              verification_state:
                const: PASSED
              target_exclusivity_status:
                const: ACTIVE
              target_copy_lock:
                properties:
                  state:
                    const: RELEASED
                  activated_at:
                    $ref: '#/components/schemas/Timestamp'
                  released_at:
                    $ref: '#/components/schemas/Timestamp'
                  fence_epoch:
                    type: integer
                    minimum: 1
              exit_code:
                const: 0
              verification_summary:
                $ref: '#/components/schemas/VerificationSummary'
            oneOf:
              - properties:
                  summary_parse_status:
                    const: SUCCEEDED
                  run_summary:
                    allOf:
                      - $ref: '#/components/schemas/RunSummary'
                      - properties:
                          records_failed:
                            const: 0
              - properties:
                  summary_parse_status:
                    const: FAILED
                  run_summary:
                    type: 'null'
                description: 日志汇总解析失败只丢失诊断指标；固定 DataX errorLimit=0、exit 0 与独立 oracle PASSED 仍可形成业务成功。
        - if:
            properties:
              process_state:
                const: VERIFYING
            required: [process_state]
          then:
            properties:
              verification_state:
                enum: [NOT_STARTED, VERIFYING]
              exit_code:
                const: 0
        - if:
            properties:
              verification_state:
                const: NOT_STARTED
            required: [verification_state]
          then:
            properties:
              verification_summary:
                type: 'null'
        - if:
            properties:
              verification_state:
                enum: [PASSED, FAILED, INCONCLUSIVE]
            required: [verification_state]
          then:
            properties:
              verification_summary:
                $ref: '#/components/schemas/VerificationSummary'
        - if:
            properties:
              verification_state:
                const: PASSED
            required: [verification_state]
          then:
            properties:
              verification_summary:
                allOf:
                  - $ref: '#/components/schemas/VerificationSummary'
                  - properties:
                      result:
                        const: PASSED
        - if:
            properties:
              verification_state:
                const: FAILED
            required: [verification_state]
          then:
            properties:
              verification_summary:
                allOf:
                  - $ref: '#/components/schemas/VerificationSummary'
                  - properties:
                      result:
                        const: FAILED
        - if:
            properties:
              verification_state:
                const: INCONCLUSIVE
            required: [verification_state]
          then:
            properties:
              verification_summary:
                allOf:
                  - $ref: '#/components/schemas/VerificationSummary'
                  - properties:
                      result:
                        const: INCONCLUSIVE
        - if:
            properties:
              verification_state:
                enum: [FAILED, INCONCLUSIVE]
            required: [verification_state]
          then:
            properties:
              process_state:
                const: FAILED
...
    ExecutionRerunCreate:
      type: object
      additionalProperties: false
      required: [source_quiescence_confirmation, target_exclusivity_confirmation, recovery_gate_id]
      properties:
        source_quiescence_confirmation:
          $ref: '#/components/schemas/SourceQuiescenceConfirmation'
        target_exclusivity_confirmation:
          $ref: '#/components/schemas/TargetExclusivityConfirmation'
        recovery_gate_id:
          $ref: '#/components/schemas/Uuid'
          description: 必须引用原执行经如实处置确认与 Worker 空表复检得到的 VERIFIED 门禁；确认可明确无需清理。
    RecoveryGate:
      type: object
      additionalProperties: false
      required:
        - id
        - execution_id
        - target_namespace_id
        - status
        - data_effect_at_open
        - remediation_confirmation
        - target_empty_evidence
        - latest_recovery_probe_id
        - submitted_at
        - verified_at
        - reason_code
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        execution_id:
          $ref: '#/components/schemas/Uuid'
        target_namespace_id:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [OPEN, REMEDIATION_SUBMITTED, VERIFIED, REJECTED]
        data_effect_at_open:
          $ref: '#/components/schemas/DataEffect'
        remediation_confirmation:
          oneOf:
            - $ref: '#/components/schemas/RemediationConfirmation'
            - type: 'null'
        target_empty_evidence:
          oneOf:
            - $ref: '#/components/schemas/TargetEmptyEvidence'
            - type: 'null'
        latest_recovery_probe_id:
          $ref: '#/components/schemas/NullableUuid'
        submitted_at:
          $ref: '#/components/schemas/NullableTimestamp'
        verified_at:
          $ref: '#/components/schemas/NullableTimestamp'
        reason_code:
          type: [string, 'null']
          maxLength: 64
...
    RecoveryProbe:
      type: object
      additionalProperties: false
      description: 不启动 DataX 的独立恢复空表探针；使用自身事实队列、Attempt、lease 和 fence。
      required:
        - id
        - project_id
        - recovery_gate_id
        - target_namespace_id
        - target_datasource_revision_id
        - target_endpoint_policy_revision_id
        - process_state
        - result
        - fence_epoch
        - active_attempt
        - service_reservation_seconds
        - queue_eligibility_state
        - queue_block_reason
        - queue_state_changed_at
        - eligible_wait_milliseconds
        - queued_at
        - finished_at
        - target_connection_evidence_id
        - target_empty_evidence
        - failure_code
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        project_id:
          $ref: '#/components/schemas/Uuid'
        recovery_gate_id:
          $ref: '#/components/schemas/Uuid'
        target_namespace_id:
          $ref: '#/components/schemas/Uuid'
        target_datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        target_endpoint_policy_revision_id:
          $ref: '#/components/schemas/NullableUuid'
        process_state:
          type: string
          enum: [QUEUED, STARTING, RUNNING, SUCCEEDED, FAILED, CANCELED, LOST]
        result:
          type: string
          enum: [NOT_STARTED, EMPTY, NONEMPTY, INCONCLUSIVE]
        fence_epoch:
          type: integer
          minimum: 0
        active_attempt:
          oneOf:
            - $ref: '#/components/schemas/RecoveryProbeAttempt'
            - type: 'null'
        service_reservation_seconds:
          type: integer
          minimum: 1
        queue_eligibility_state:
          type: string
          enum: [ELIGIBLE, BLOCKED]
        queue_block_reason:
          type: [string, 'null']
          maxLength: 64
        queue_state_changed_at:
          $ref: '#/components/schemas/Timestamp'
        eligible_wait_milliseconds:
          type: integer
          minimum: 0
        queued_at:
          $ref: '#/components/schemas/Timestamp'
        finished_at:
          $ref: '#/components/schemas/NullableTimestamp'
        target_connection_evidence_id:
          $ref: '#/components/schemas/NullableUuid'
        target_empty_evidence:
          oneOf:
            - $ref: '#/components/schemas/TargetEmptyEvidence'
            - type: 'null'
        failure_code:
          type: [string, 'null']
          maxLength: 64
      allOf:
        - if:
            properties:
              process_state:
                const: SUCCEEDED
            required: [process_state]
          then:
            properties:
              result:
                const: EMPTY
    RecoverySubmission:
      type: object
      additionalProperties: false
      required: [recovery_gate, recovery_probe]
      properties:
        recovery_gate:
          $ref: '#/components/schemas/RecoveryGate'
        recovery_probe:
          $ref: '#/components/schemas/RecoveryProbe'
...